Update index.html
During my review of your HTML and JavaScript code, I noticed an issue in the section where you handle the editing of a reservation. Specifically, the problem was in the part of the code where I create the updatedReservation object.

In the original code, there was a syntax error in the updatedReservation object creation. The roomType property was not properly assigned a value. The corrected line of code should look like this:
roomType: $('#roomType').val(),